refactored content editor tabs (#2881)
